Use platformdir for config dirs, and restructure
