+++ pyproject.toml
@@ -37,6 +37,7 @@ dependencies = [
     "pyyaml",
     "pyperclip",
     "dateparser",
+    "platformdirs",
     "urllib3",
     "orjson",
     "importlib-metadata >= 1.0; python_version < \"3.8\"",
